% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lereanalisartsmaiscompleta.R
\name{lereanalisartsmaiscompleta}
\alias{lereanalisartsmaiscompleta}
\title{Função para criação de série temporal com valores mensais a partir da leitura de arquivo em csv que possua cabeçalho com campos separados por vírgula com ano, mês e valor. Além disso, apresenta resultados de sua análise exploratória mais completa.}
\usage{
lereanalisartsmaiscompleta(a, m)
}
\arguments{
\item{a}{Valor do ano que inicia a série temporal}

\item{m}{Valor do mês que inicia a série temporal}
}
\description{
Esta função serve para criar um série temporal com valores mensais e analisá-la estatisticamente mais completa
}
\examples{
lereanalisartsmaiscompleta(2001,1)

}
